De-JUCE: AudioEngine MIDI plumbing onto the duskstudio::midi layer
Deletes the engine-side device banks, input callback, out-FIFO and pump
(now owned by MidiDevices) and unifies the MIDI scratches on
dusk::MidiBuffer: perInputMidi drains through MidiInputClient, the
sync/MCU/MTC receiver bridges collapse to direct buffer passes, and the
MTC emitter writes into the shared clock-out scratch. Bridges kept where
consumers stay JUCE: per-track routing (PluginSlot + RecordManager) and
the test-inject merge.

MidiClockEmitter flips to dusk::MidiBuffer raw-byte output and leaves
the allowlist with its cpp (net tower ratchet zero: the seam layer's
two files replace them). UI/serializer size() casts follow the new
device-info vector; MCU sink sends through the bank's JUCE overload.

364/364 tests (clock drift suite rewritten dusk-side), selftest 26
PASS/0 FAIL under Xvfb with live device enumeration, gate 202.

+++ src/DuskStudioApp.cpp
@@ -553,7 +553,7 @@ static void runHeadlessPipelineTest (const juce::String& pluginPath)
     // - Otherwise pick index 0 if any MIDI input exists, and force track
     //   0's midiInputIndex to match so the test can inject end-to-end.
     int midiInputIdx = -1;
-    const int numMidiInputs = engine->getMidiInputDevices().size();
+    const int numMidiInputs = (int) engine->getMidiInputDevices().size();
     if (numMidiInputs == 0)
     {
         std::fprintf (stderr, "WARN: no MIDI inputs available; injection will be a no-op.\n");
